pyproject.toml
[tool.poetry]
name = "medical-diagnostic-center"
version = "0.1.0"
description = ""
authors = ["Евгений Ермак <144560375+EvgeniiErmak@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^5.0.4"
psycopg2-binary = "^2.9.9"
django-crispy-forms = "^2.1"
pillow = "^10.3.0"
djangorestframework = "^3.15.1"
channels = "^4.1.0"
django-allauth = "^0.61.1"
psycopg2 = "^2.9.9"
django-bootstrap4 = "^24.1"
django-storages = "^1.14.2"
django-ckeditor-5 = "^0.2.12"
django-summernote = "^0.8.20.0"
django-countries = "^7.6.1"
aiortc = "^1.8.0"
channels-redis = "^4.2.0"
drf-yasg = "^1.21.7"
sphinx = "^7.3.0"
sphinx-rtd-theme = "^2.0.0"
sphinx-autobuild = "^2024.4.16"
python-decouple = "^3.8"
python-dotenv = "^1.0.1"
django-environ = "^0.11.2"


[tool.poetry.group.dev.dependencies]
flake8 = "^7.0.0"
autopep8 = "^2.1.0"
black = "^24.4.0"
yapf = "^0.40.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
